fix : getter 수정 버그 찾기

+++ src/main/java/domain/LottoResult.java
@@ -14,7 +14,7 @@ public class LottoResult {
     }
 
     public double profitPercent(int purchaseMoney) {
-        return (double) allWinning() / purchaseMoney * 100;
+        return  ((double)allWinning() / (double)purchaseMoney) * 100;
     }
 
     private long allWinning() {
